Working...

+++ java/src/main/java/roulette/RouletteService.java
@@ -3,7 +3,7 @@ package roulette;
 public class RouletteService {
 
 	public int getRandomNumber() {
-		return Math.random();
+		return (int) Math.random();
 	}
 
 }
